ai-auto-applier/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
psycopg2-binary
python-dotenv